requirements.txt
asgiref==3.7.2
asttokens==2.4.1
bcrypt==3.2.0
black==24.2.0
blinker==1.4
build==1.0.3
certifi==2020.6.20
cffi==1.16.0
chardet==4.0.0
cliapp==1.0.9
click==8.0.3
cryptography==3.4.8
decorator==5.1.1
defusedxml==0.8.0rc2
distro==1.7.0
Django==5.0.3
django-cors-headers==4.4.0
djangorestframework==3.15.2
docutils==0.20.1
exceptiongroup==1.2.0
executing==2.0.1
Faker==8.12.1
fasteners==0.14.1
future==0.18.2
httplib2==0.20.2
idna==2.5
importlib-metadata==4.6.4
iniconfig==2.0.0
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.1
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.3.6
markdown-it-py==3.0.0
MarkupSafe==2.0.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mock==5.1.0
monotonic==1.6
more-itertools==8.10.0
mypy-extensions==1.0.0
netifaces==0.11.0
nh3==0.2.15
oauthlib==3.2.0
olefile==0.46
packaging==23.2
paramiko==2.9.3
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
Pillow==9.0.1
pkginfo==1.9.6
platformdirs==4.2.0
pluggy==1.3.0
prompt-toolkit==3.0.41
protobuf==3.12.4
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.22
Pygments==2.17.2
PyJWT==2.8.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyproject_hooks==1.0.0
PyQt5==5.15.6
PyQt5-Qt5==5.15.14
PyQt5-sip==12.9.1
pyRFC3339==1.1
python-dateutil==2.8.1
python-decouple==3.8
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2022.1
pyxattr==0.7.2
pyxdg==0.27
readme-renderer==42.0
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.7.0
SecretStorage==3.3.1
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.4.4
stack-data==0.6.3
text-unidecode==1.3
tomli==2.0.1
tornado==6.1
traitlets==5.14.0
twine==5.0.0
typing_extensions==4.10.0
urllib3==1.26.5
wadllib==1.3.6
wcwidth==0.2.12
xdg==5.0.0
xlrd==1.2.0
zipp==1.0.0
